Refactor and added csv batch add
